requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.2
pymysql==1.0.2
sqlalchemy==1.4.25
flask-sqlalchemy==2.5.1